fix ios8 build errors and warnings; 64 and 32 bit issues, obj_msgSend casting, delegate casting

+++ Janrain/JREngage/Classes/JREngageError.m
@@ -36,7 +36,7 @@ NSString *JREngageErrorDomain = @"JREngage.ErrorDomain";
 @implementation JREngageError
 + (NSError *)errorWithMessage:(NSString *)message andCode:(NSInteger)code
 {
-    ALog (@"An error occurred (%d): %@", code, message);
+    ALog (@"An error occurred (%ld): %@", (long)code, message);
     NSDictionary *userInfo = [NSDictionary dictionaryWithObjectsAndKeys:
                                                    [NSString stringWithString:message], NSLocalizedDescriptionKey, nil];
